requirements.txt
mysql-connector-python
PyYAML
requests
pycountry
feedparser
beautifulsoup4
yfinance